cc: codegen: move expression handling to separate file
