+++ src/cc/codegen.cpp
@@ -6,50 +6,6 @@
 using namespace std;
 using namespace antlr4;
 
-ConstantValue parse_constant(string text)
-{
-    // Note: I'm not handling exceptions here, because I trust the ANTLR parser to not let anything weird through.
-    if (isdigit(text[0]))
-    {
-        if (text[0] != '0' || text.size() == 1)
-        {
-            if (text.find('.') != string::npos)
-            {
-                return stof(text);
-            }
-            else
-            {
-                return stoi(text);
-            }
-        }
-        else
-        {
-            char base = tolower(text[1]);
-            if (base == 'x' || base == 'X')
-            {
-                return stoi(text, nullptr, 16);
-            }
-            else if (base == 'b' || base == 'B')
-            {
-                return stoi(text, nullptr, 2);
-            }
-            else
-            {
-                return stoi(text, nullptr, 8);
-            }
-        }
-    }
-    else if (text[0] == '\'')
-    {
-        // TODO: Unescape characters
-        return text[1];
-    }
-    else
-    {
-        throw runtime_error("unhandled constant type");
-    }
-}
-
 CodeGen::CodeGen(ProgramMeta &program_meta, std::ostream &output) : program_meta(program_meta), output(output)
 {
     output << "\t; Generated by ZDK C Compiler\n\n";
@@ -128,21 +84,4 @@ any CodeGen::visitJumpStatement(CParser::JumpStatementContext *ctx)
     }
 
     return any();
-}
-
-any CodeGen::visitPrimaryExpression(CParser::PrimaryExpressionContext *ctx)
-{
-    ExpressionCtx expr_ctx;
-
-    if (auto const_ctx = ctx->Constant())
-    {
-        expr_ctx.constant = true;
-        expr_ctx.value = parse_constant(const_ctx->getText());
-    }
-    else
-    {
-        throw runtime_error("unsupported expression type");
-    }
-
-    return expr_ctx;
 }
